Truth Grounder AI AI-Powered Fact Verification and Correction System/requirements.txt
transformers
sentence-transformers
faiss-cpu
spacy
torch
